requirements/base.txt
django==3.2.5
djangorestframework==3.12.4
hubspot-api-client==3.8.2